docs: introduce documentation hosted on GitHub pages (#18)
**Issue #, if available:**

## Description of changes:

To make it easier to use the module user documentation is needed. By
offering this through GitHub pages we have a nice interactive website
that helps you with that.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [x] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.