fix: homepage url in package and links within documentation (#19)
**Issue #, if available:**

## Description of changes:

The package on pipit did not include the proper links to the bug
trackers and user documentation. Also some of the Hugo redirects where
not functioning as they should so fixed that by applying a workaround.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [x] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.