fix: correctly link to user documentation (#20)
**Issue #, if available:**

## Description of changes:

Link to the homepage still did not resolve correctly with the previous
fix. Renamed the `Usage` section to `User Documentation` and added a
direct link to the `User Documentation`.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [x] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.