feat: generate profiles for aws cli tool (#22)
**Issue #, if available:**

## Description of changes:

Managing profiles for the AWS cli could become a nightmare when you have
a lot of accounts. To make it easier you can generate a separate config
file for your organization.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [x] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.