feat: return environments in DTAP order (#85)
**Issue #, if available:**

## Description of changes:

When you retrieve the environments you used to get a random order. With
this feature you will have the ability to customize the order. By
default we follow the DTAP order as in Development, Testing, Acceptance
and Production.

**Checklist**

<!--- Leave unchecked if your change doesn't seem to apply -->

* [x] Update tests
* [x] Update docs
* [x] PR title follows [conventional commit
semantics](https://www.conventionalcommits.org/en/v1.0.0-beta.2/#commit-message-for-a-fix-using-an-optional-issue-number)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.